+++ src/Commands/ContentTypes/AddContentTypesFromContentTypeHub.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.Management.Automation;
 using Microsoft.SharePoint.Client;
-
 using PnP.PowerShell.Commands.Base.PipeBinds;
 
 namespace PnP.PowerShell.Commands.ContentTypes
@@ -13,21 +12,29 @@ namespace PnP.PowerShell.Commands.ContentTypes
         [ValidateNotNullOrEmpty]
         public List<string> ContentTypes;
 
+        [Parameter(Mandatory = false)]
+        public SitePipeBind Site;
+
         protected override void ExecuteCmdlet()
         {
-            Microsoft.SharePoint.Client.Site site = ClientContext.Site;
-            ClientContext.Load(site);
-            ClientContext.ExecuteQueryRetry();
+            var site = ParameterSpecified(nameof(Site)) ? Site.Url : ClientContext.Url;
 
             var sub = new Microsoft.SharePoint.Client.Taxonomy.ContentTypeSync.ContentTypeSubscriber(ClientContext);
             ClientContext.Load(sub);
             ClientContext.ExecuteQueryRetry();
 
-            var res = sub.SyncContentTypesFromHubSite2(site.Url, ContentTypes);
+            var res = sub.SyncContentTypesFromHubSite2(ClientContext.Url, ContentTypes);
             ClientContext.ExecuteQueryRetry();
-            
-            WriteObject(res);
-        }
 
+            var result = new PnP.PowerShell.Commands.Model.SharePoint.AddContentTypesFromContentTypeHubResponse
+            {
+                FailedContentTypeErrors = res.Value.FailedContentTypeErrors,
+                FailedReason = res.Value.FailedReason,
+                FailedContentTypeIDs = res.Value.FailedContentTypeIDs,
+                IsPassed = res.Value.IsPassed,
+                Value = res.Value
+            };
+            WriteObject(result);
+        }
     }
 }